feat: add basic structure
