+++ package/lib/comprehension_measurement.dart
@@ -1,28 +1,2 @@
-library comprehension_measurement;
-
-import 'package:flutter/material.dart';
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
-
-void measureComprehension(BuildContext context) {
-  showBottomSheet(
-    context: context,
-    backgroundColor: Colors.transparent,
-    builder: (BuildContext context) {
-      return Container(
-        margin: const EdgeInsets.all(8),
-        width: double.infinity,
-        child: Card(
-          elevation: 4,
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-          child: const Text("Text"),
-        ),
-      );
-    },
-  );
-}
+export 'src/main.dart' show measureComprehension;
+export 'src/types/single_choice.dart' show SingleChoiceWidget;
