+++ package/lib/src/main.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+void measureComprehension(BuildContext context) {
+  showBottomSheet(
+    context: context,
+    backgroundColor: Colors.transparent,
+    builder: (BuildContext context) {
+      return Container(
+        margin: const EdgeInsets.all(8),
+        width: double.infinity,
+        child: Card(
+          elevation: 4,
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+          child: const Text("Text"),
+        ),
+      );
+    },
+  );
+}
