+++ package/lib/src/types/single_choice.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class SingleChoiceWidget extends StatefulWidget {
+  const SingleChoiceWidget({Key? key}) : super(key: key);
+
+  @override
+  State<SingleChoiceWidget> createState() => _SingleChoiceWidgetState();
+}
+
+class _SingleChoiceWidgetState extends State<SingleChoiceWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
