+++ package/test/comprehension_measurement_test.dart
@@ -3,10 +3,5 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:comprehension_measurement/comprehension_measurement.dart';
 
 void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
-  });
+  // TODO: Add tests
 }
